incorporated reference declarations fully
by bypassing all safety procedures :p. not really though, there's a
reference safety proc now that scans to see if references are valid and
if so it expands the environment using them. next step is to make the
environment expander recognize reference declarations and treat them in
a specific fashion.