debugging + important change for environment (READ)
since all lets and whatnot get defaulted into lambdas, all closures
created CANNOT have access to variables in their applications calls, but
only values. this, in part, is due to the fact that we implement
environment lookups in a bottom -> top approach. how do we solve this?
whenever there is a procedure application, we pass in the current
environment at the procedure call. this way, for variables that need to
be called by reference, we can lookup addresses. then we extend the
environment like we normally would with closure application.